+++ src/main/webapp/wise5/components/conceptMap/conceptMapAuthoringComponentModule.js
@@ -1,0 +1,26 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+
+var _conceptMapService = require('./conceptMapService');
+
+var _conceptMapService2 = _interopRequireDefault(_conceptMapService);
+
+var _conceptMapController = require('./conceptMapController');
+
+var _conceptMapController2 = _interopRequireDefault(_conceptMapController);
+
+var _conceptMapAuthoringController = require('./conceptMapAuthoringController');
+
+var _conceptMapAuthoringController2 = _interopRequireDefault(_conceptMapAuthoringController);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+var conceptMapAuthoringComponentModule = angular.module('conceptMapAuthoringComponentModule', ['pascalprecht.translate']).service(_conceptMapService2.default.name, _conceptMapService2.default).controller(_conceptMapController2.default.name, _conceptMapController2.default).controller(_conceptMapAuthoringController2.default.name, _conceptMapAuthoringController2.default).config(['$translatePartialLoaderProvider', function ($translatePartialLoaderProvider) {
+  $translatePartialLoaderProvider.addPart('components/conceptMap/i18n');
+}]);
+
+exports.default = conceptMapAuthoringComponentModule;
+//# sourceMappingURL=conceptMapAuthoringComponentModule.js.map
