+++ src/main/webapp/wise5/directives/component/component.es6
@@ -80,7 +80,7 @@ class ComponentController {
 
         }
 
-        if ($scope.mode === 'authoring' && componentContent.type == 'HTML') {
+        if ($scope.mode === 'authoring' && (componentContent.type == 'HTML' || componentContent.type == 'ConceptMap')) {
             $scope.componentTemplatePath = this.NodeService.getComponentAuthoringTemplatePath(componentContent.type);
         } else {
             $scope.componentTemplatePath = this.NodeService.getComponentTemplatePath(componentContent.type);
